API: rm arg `depth` from `ast.Operator.__str__`
Passing this argument to operands raises errors with
AST nodes that are instances of classes not designed
for parsing, for example `list`. Such a limitation is
inconvenient and unnecessary.

In 3 years, I have never used the argument `depth`.
Given the rarity of this operation, how frequently
it raises errors, and the complexity it can induce,
instead use a depth-first search function to trim the
AST prior to calling `__str__`.