BUG: pass `debuglog` to `yacc` whenever `debug`
because the function `ply.yacc.yacc()` does not close the
logging file opened when it instantiates a `PlyLogger`.
The result of this is an unclosed file.

As a result, a `ResourceWarning` (["Warnings"][ResourceWarning_docs],
["What's New In Python 3.2"][py32], ["ResourceWarning Example"][devmode])
is issued by Python:

> ResourceWarning: unclosed file
> `<_io.TextIOWrapper name='./parser.out' mode='w' encoding='UTF-8'>`

This refers to `ply == 3.10`.
The default name of this file in `ply` is `parser.out`.

The warning type `ResourceWarning` was introduced in Python 3.2,
is not printed by default, and can be enabled using `python -X dev`.
The [documentation reads][py32]:

> A [`ResourceWarning`][ResourceWarning_docs] is also issued
> when a [`file object`][glossary_file_object] is destroyed without
> having been explicitly closed.
> While the deallocator for such object ensures it closes the
> underlying operating system resource (usually, a file descriptor),
> the delay in deallocating the object could produce various issues,
> especially under Windows.

Therefore the file is closed by Python,
but the file needs to be closed explicitly.

The file object is stored in an attribute of an instance of
the class `ply.yacc.PlyLogger`.

`ply.yacc.yacc()` uses as default value for the
parameter `debugfile` the module variable `debug_file`,
which is initialized to the default value `'parser.out'`.


[ResourceWarning_docs]: https://docs.python.org/3/library/exceptions.html#ResourceWarning
[py32]: https://docs.python.org/3/whatsnew/3.2.html#other-language-changes
[devmode]: https://docs.python.org/3/library/devmode.html#resourcewarning-example
[glossary_file_object]: https://docs.python.org/3/glossary.html#term-file-object