MNT: use raw strings where `\` appears
This avoids a `DeprecationWarning` that was introduced
for this syntax in Python 3.6, and is expected to
become a `SyntaxError` in a future Python version.

Quoting from ["Deprecated Python behavior"][dpy] in What’s New In Python 3.6:

> A backslash-character pair that is not a valid escape sequence now generates
> a `DeprecationWarning`. Although this will eventually become a `SyntaxError`,
> that will not be for several Python releases.

Read also [bpo-27364][bpo27364] and [bpo-32912][bpo32912].

Quoting from ["String and Bytes literals"][str_bytes] in the Python documentation:

> Changed in version 3.6: Unrecognized escape sequences produce
> a `DeprecationWarning`. In a future Python version they will be
> a `SyntaxWarning` and eventually a `SyntaxError`.

This `DeprecationWarning` can be printed by using `python -X dev`.

[dpy]: https://docs.python.org/3/whatsnew/3.6.html#deprecated-python-behavior
[bpo27364]: https://bugs.python.org/issue27364
[bpo32912]: https://bugs.python.org/issue32912
[str_bytes]: https://docs.python.org/3.10/reference/lexical_analysis.html#string-and-bytes-literals