TST: use `pytest`, instead of `nose`
because `nose` is not compatible with Python 3.10.

- REL: require `pytest >= 4.6.11` in `tests_require`

On Travis CI running `nosetests` previously succeeded,
without `nose` being installed by the package or `.travis.yml`,
nor `nose` appearing anywhere in the log of Travis CI.

[1]: https://docs.travis-ci.com/user/reference/trusty/#pre-installed-pip-packages

Coverage is not any more computed in CI runs,
because on GitHub a miscellaneous reduction in coverage
can result in the build status of a commit displaying as failed.

This could give the impression that the tests failed,
when it was only a change of coverage that caused
the displayed status.