BUG: list tokens in same order in different runs
because otherwise `ply` can produce a parser
signature different than the signature that
`ply` computed when caching the LALR(1) state
machine of the same parser.

When this change of signatures occurs,
`ply` rebuilds the LALR(1) state machine,
even though the lexer and parser grammars
have not changed since the last time that
the LALR(1) state machine was cached.

By sorting the token names, this issue is avoided.

The revised method `astutils.ply.Parser.__init__()`
calls `sorted()` on the attribute `self.reserved`.
This attribute is by default a dictionary, and so
assumed that it is dictionary-like.

Using `sorted()` ensures LALR(1) machine signature
independent of the order in which the token names
are written in Python source code.