DEP: do not implicitly set attribute `Parser.Lexer` to class `Lexer`
Explicit is better than implicit.
So now the parser's lexer can be defined in two ways:

- define a new class with an attribute `Lexer` (this approach
  is about to be deprecated)

- pass the argument `lexer` to the method
  `astutils.ply.Parser.__init__()`

The approach that is removed by this commit was
likely unused, because `astutils.ply.Lexer`
has no grammar-equalities defined. So this class
could not have been used for parsing.
Thus, this change is expected to not affect uses of
the class `astutils.ply.Parser`.